Improved Garage Scene and implemented basic UI
