+++ Assets/Imported Assets/Garage.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d7c954c7ee45406a9e482e2c58b9990
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
