+++ Assets/Imported Assets/Garage/scifigarage.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 73d41bc9c77da48068c7fd52d8d83a39
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
