+++ Assets/Scripts/Features/Gameplay/Cosmetics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 26b9b2be8e26f4109836ce6001086cd9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
