+++ Assets/Scripts/Features/Gameplay/Cosmetics/CosmeticsManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a81468b55c487422ab106cfdf888d5a6
