+++ Assets/Scripts/Features/UI/CosmeticsButton.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3b1cc18e64ed54a3a909edb0d8652cce
