+++ Assets/Scripts/Features/UI/ShipRotation.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cd9c8c1926a2e4f53bc80379ed88987d
